refactor : 디렉터리 구조 변경
Related to : #17
