+++ src/main/java/com/umc5th/muffler/alarm/service/AlarmService.java
@@ -1,9 +1,8 @@
 package com.umc5th.muffler.alarm.service;
 
-import com.umc5th.muffler.alarm.dto.Alarmable;
 import java.util.List;
 
 public interface AlarmService {
     void sendAlarm(Alarmable alarmable);
-    <T extends Alarmable> void sendAlarms(List<T> alarmables);
+    <T extends Alarmable> void sendAlarms(List<T> alarms);
 }
