+++ src/main/java/com/umc5th/muffler/domain/dailyplan/repository/DailyPlanRepositoryImpl.java
@@ -1,8 +1,12 @@
 package com.umc5th.muffler.domain.dailyplan.repository;
 
 import com.querydsl.core.Tuple;
+import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.umc5th.muffler.domain.dailyplan.dto.DailyPlanAlarm;
 import com.umc5th.muffler.entity.QDailyPlan;
+import com.umc5th.muffler.entity.QMember;
+import com.umc5th.muffler.entity.QMemberAlarm;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -22,5 +26,26 @@ public class DailyPlanRepositoryImpl implements DailyPlanRepositoryCustom {
                 .where(dailyPlan.goal.id.eq(goalId)
                         .and(dailyPlan.date.between(startDate, endDate))
                 ).fetch();
+    }
+
+    @Override
+    public List<DailyPlanAlarm> findDailyPlanAlarms(LocalDate date) {
+        QMemberAlarm memberAlarm =  QMemberAlarm.memberAlarm;
+        QMember member = QMember.member;
+        QDailyPlan dailyPlan = QDailyPlan.dailyPlan;
+
+        return queryFactory
+                .select(Projections.fields(DailyPlanAlarm.class,
+                        member.name.as("memberName"),
+                        dailyPlan.budget.as("dailyBudget"),
+                        memberAlarm.token.as("alarmToken")))
+                .from(member)
+                .join(memberAlarm).on(memberAlarm.member.id.eq(member.id))
+                .join(dailyPlan).on(dailyPlan.goal.member.id.eq(member.id))
+                .where(
+                        dailyPlan.date.eq(date),
+                        memberAlarm.isDailyPlanRemindAgree.eq(true)
+                )
+                .fetch();
     }
 }
