+++ src/main/java/com/umc5th/muffler/schedule/service/ScheduledAlarmService.java
@@ -1,9 +1,9 @@
 package com.umc5th.muffler.schedule.service;
 
 import com.umc5th.muffler.alarm.service.AlarmService;
+import com.umc5th.muffler.domain.dailyplan.repository.DailyPlanRepository;
 import com.umc5th.muffler.global.util.DateTimeProvider;
-import com.umc5th.muffler.schedule.repository.AlarmRepository;
-import com.umc5th.muffler.alarm.dto.DailyPlanAlarm;
+import com.umc5th.muffler.domain.dailyplan.dto.DailyPlanAlarm;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -15,14 +15,14 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ScheduledAlarmService {
     private final DateTimeProvider dateTimeProvider;
-    private final AlarmRepository alarmRepository;
+    private final DailyPlanRepository dailyPlanRepository;
     private final AlarmService alarmService;
 
     @Transactional(readOnly = true)
     @Scheduled(cron = "0 0 9 * * *", zone = "Asia/Seoul")
     public void alarmTodayPlan() {
         LocalDate today = dateTimeProvider.nowDate();
-        List<DailyPlanAlarm> alarms = alarmRepository.findDailyPlanAlarms(today);
+        List<DailyPlanAlarm> alarms = dailyPlanRepository.findDailyPlanAlarms(today);
         alarmService.sendAlarms(alarms);
     }
 }
