+++ src/main/resources/application.yml
@@ -79,3 +79,4 @@ server:
       force: true
 logging.level:
   org.hibernate.SQL: debug
+  org.hibernate.type.descriptor.sql: trace
